+++ starbear/__init__.py
@@ -2,7 +2,7 @@ from hrepr import H
 
 from .constructors import BrowserEvent, FormData, NamespaceDict
 from .page import Component, JavaScriptOperation, Page, selector_for
-from .ref import Reference
+from .reg import Reference
 from .serve import (
     ConfigurableBear,
     ConfigurableSimpleBear,
